requirements.txt
django == 3.2
djangorestframework == 3.12
djongo == 1.3.6
dnspython3 == 1.15.0
drf-yasg == 1.20.0
ipdb == 0.13.9
